Refactor Remix doctor CLI invocation
